refactor(chat): extract chat_manager/send_entry.rs (#292)
Move send_message, set_working, edit_queued_message,
cancel_queued_message, and find_ref off the ChatManager facade —
pure move, no behavior change. set_working becomes pub(super) for
send.rs's existing callers.
